LoongArch: testsuite: refine __tls_get_addr tests with tls_native
If GCC is not built with a working linker for the target (developers
occansionally build such a "minimal" GCC for testing and debugging),
TLS will be emulated and __tls_get_addr won't be used.  Refine those
tests depending on __tls_get_addr with tls_native to avoid test
failures.

gcc/testsuite/ChangeLog:

	* gcc.target/loongarch/func-call-medium-1.c: Refine test
	depending on __tls_get_addr with { target tls_native }.
	* gcc.target/loongarch/func-call-medium-2.c: Likewise.
	* gcc.target/loongarch/func-call-medium-3.c: Likewise.
	* gcc.target/loongarch/func-call-medium-4.c: Likewise.
	* gcc.target/loongarch/func-call-medium-5.c: Likewise.
	* gcc.target/loongarch/func-call-medium-6.c: Likewise.
	* gcc.target/loongarch/func-call-medium-7.c: Likewise.
	* gcc.target/loongarch/func-call-medium-8.c: Likewise.
	* gcc.target/loongarch/tls-gd-noplt.c: Likewise.